+++ l10n_es_facturae/models/account_move.py
@@ -157,8 +157,6 @@ class AccountMove(models.Model):
             raise ValidationError(_("Partner vat is too small"))
         if not self.partner_id.state_id:
             raise ValidationError(_("Partner state not provided"))
-        if not self.partner_id.organo_gestor:
-            raise ValidationError(_("Organo Gestor not provided"))
         if not self.partner_id.unidad_tramitadora:
             raise ValidationError(_("Unidad Tramitadora not provided"))
         if not self.partner_id.oficina_contable:
